+++ crates/icbsetup/src/editors/door.rs
@@ -267,6 +267,7 @@ impl<'a> Page for DoorEditor<'a> {
                             lock.push(Door {
                                 name,
                                 number: 0,
+                                valid: false,
                                 description: "".to_string(),
                                 password: "".to_string(),
                                 securiy_level: SecurityExpression::default(),