+++ crates/icy_board_engine/src/icy_board/doors/mod.rs
@@ -121,6 +121,9 @@ pub struct Door {
     #[serde(skip)]
     pub number: usize,
 
+    #[serde(skip)]
+    pub valid: bool,
+
     #[serde(default)]
     #[serde(skip_serializing_if = "SecurityExpression::is_empty")]
     #[serde_as(as = "DisplayFromStr")]
@@ -160,6 +163,7 @@ impl UserData for Door {
     fn register_members<F: UserDataMemberRegistry>(registry: &mut F) {
         registry.add_property(NAME.clone(), VariableType::String, false);
         registry.add_property(NUMBER.clone(), VariableType::Integer, false);
+        registry.add_property(VALID.clone(), VariableType::Boolean, false);
         registry.add_property(DESCRIPTION.clone(), VariableType::String, false);
         registry.add_property(PASSWORD.clone(), VariableType::Password, false);
         registry.add_function(HAS_ACCESS.clone(), Vec::new(), VariableType::Boolean);
@@ -174,6 +178,9 @@ impl UserDataValue for Door {
         }
         if *name == *NUMBER {
             return Ok(VariableValue::new_int(self.number as i32));
+        }
+        if *name == *VALID {
+            return Ok(VariableValue::new_bool(self.valid));
         }
         if *name == *DESCRIPTION {
             return Ok(VariableValue::new_string(self.description.clone()));
@@ -214,6 +221,7 @@ impl UserDataValue for Door {
 
 pub static NAME: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("Name".to_string()));
 pub static NUMBER: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("Number".to_string()));
+pub static VALID: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("Valid".to_string()));
 pub static DESCRIPTION: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("Description".to_string()));
 pub static PASSWORD: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("Password".to_string()));
 pub static HAS_ACCESS: std::sync::LazyLock<unicase::Ascii<String>> = std::sync::LazyLock::new(|| unicase::Ascii::new("HasAccess".to_string()));
@@ -252,6 +260,7 @@ impl DoorList {
 
             let door = Door {
                 number: 0,
+                valid: false,
                 name: file.to_string(),
                 description: file.to_string(),
                 password: password.to_string(),